renames ticket to tickets
